+++ app/models/trade_log_entry.rb
@@ -1,5 +1,5 @@
 class TradeLogEntry < LogEntry
-  has_one :traded_magic_item, inverse_of: :trade_log_entry, class_name: MagicItem, dependent: :nullify
+  has_one :traded_magic_item, inverse_of: :trade_log_entry, class_name: 'MagicItem', dependent: :nullify
 
   def user
     character.user
@@ -13,7 +13,7 @@ class TradeLogEntry < LogEntry
     magic_items.count - (traded_magic_item.nil? ? 0 : 1)
   end
 
-  def magic_items_list(char)
+  def magic_items_list(char, opts={})
     traded_magic_item_name   = traded_magic_item.name if traded_magic_item
     received_magic_item      = magic_items.first
     received_magic_item_name = received_magic_item ? received_magic_item.name : ''